+++ packages/operators/src/fetch/concurrentDownload.js
@@ -1,7 +1,7 @@
 import { mergeMap, of } from 'rxjs';
 
-import { download } from './download';
+import { request } from './request';
 
 export const concurrentDownload = (concurrent = 1) => {
-  return source => source.pipe(mergeMap(url => of(url).pipe(download()), concurrent));
+  return source => source.pipe(mergeMap(url => of(url).pipe(request()), concurrent));
 };