+++ packages/operators/src/fetch/request.js
@@ -1,5 +1,6 @@
 import { concatMap } from 'rxjs';
 
+import { resolveBlob, resolveJSON, resolveText } from './response';
 import { networkRetry } from './retry';
 
 export const request = () => {
@@ -8,4 +9,16 @@ export const request = () => {
       concatMap(req => fetch(req)),
       networkRetry()
     );
+};
+
+export const requestJSON = () => {
+  return source => source.pipe(request(), resolveJSON());
+};
+
+export const requestText = () => {
+  return source => source.pipe(request(), resolveText());
+};
+
+export const requestBlob = () => {
+  return source => source.pipe(request(), resolveBlob());
 };